+++ app/src/features/apiClient/screens/apiClient/components/sidebar/APIClientSidebar.tsx
@@ -5,10 +5,10 @@ import PATHS from "config/constants/sub/paths";
 import { Tabs, TabsProps, Tooltip } from "antd";
 import { CgStack } from "@react-icons/all-files/cg/CgStack";
 import { MdOutlineHistory } from "@react-icons/all-files/md/MdOutlineHistory";
-import { CollectionsList } from "./collectionsList/CollectionsList";
+import { CollectionsList } from "./components/collectionsList/CollectionsList";
 import { MdHorizontalSplit } from "@react-icons/all-files/md/MdHorizontalSplit";
-import { HistoryList } from "./historyList/HistoryList";
-import { ApiClientSidebarHeader } from "./apiClientSidebarHeader/ApiClientSidebarHeader";
+import { HistoryList } from "./components/historyList/HistoryList";
+import { ApiClientSidebarHeader } from "./components/apiClientSidebarHeader/ApiClientSidebarHeader";
 import useEnvironmentManager from "backend/environment/hooks/useEnvironmentManager";
 import { EnvironmentsList } from "../../../environment/components/environmentsList/EnvironmentsList";
 import { useSelector } from "react-redux";
