+++ app/src/features/apiClient/screens/apiClient/components/sidebar/components/apiClientSidebarHeader/components/environmentSwitcher/environmentSwitcher.scss
