+++ app/src/features/apiClient/screens/apiClient/components/sidebar/components/collectionsList/newRecordNameInput/newRecordNameInput.scss
