+++ app/src/features/apiClient/screens/apiClient/components/sidebar/components/collectionsList/requestRow/RequestRow.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Typography } from "antd";
 import PATHS from "config/constants/sub/paths";
-import { REQUEST_METHOD_COLORS } from "../../../../../../../../constants";
+import { REQUEST_METHOD_COLORS } from "../../../../../../../../../constants";
 import { RQAPI } from "features/apiClient/types";
 import { NavLink } from "react-router-dom";
