+++ app/src/features/apiClient/screens/environment/components/environmentsList/EnvironmentsList.tsx
@@ -10,7 +10,7 @@ import PATHS from "config/constants/sub/paths";
 import { trackCreateEnvironmentClicked, trackEnvironmentCreated } from "../../analytics";
 import { actions } from "store";
 import APP_CONSTANTS from "config/constants";
-import { EmptyState } from "features/apiClient/screens/apiClient/components/sidebar/emptyState/EmptyState";
+import { EmptyState } from "features/apiClient/screens/apiClient/components/sidebar/components/emptyState/EmptyState";
 import { ListEmptySearchView } from "features/apiClient/screens/apiClient/components/sidebar/components/listEmptySearchView/ListEmptySearchView";
 import { EnvironmentAnalyticsSource } from "../../types";
 import { EnvironmentsListItem } from "./components/environmentsListItem/EnvironmentsListItem";
